新增楼层信息

+++ SunacCoordination/SunacCoordination/Sunac_def.h
@@ -112,7 +112,7 @@ public:
 	bool SetFloors(CString p_sFloors);
 	CString GetFloors()const { return m_sFloors; }
 
-	UINT GetFloorCount()const{ return m_nAllFloors.size(); }
+	int GetFloorCount()const{ return (int)(m_nAllFloors.size()); }
 	vector<int> GetAllFloor()const{ return m_nAllFloors; }
 
 	void SetFloorHeight(double p_height) { m_floorHeight = p_height; }
